Role Management System

+++ resources/views/admin/budget-usage/index.blade.php
@@ -13,7 +13,9 @@
                 <div class="card">
                     <div class="card-header">
                         <i class="fa fa-align-justify"></i> {{ trans('admin.budget-usage.actions.index') }}
-                        <a class="btn btn-primary btn-spinner btn-sm pull-right m-b-0" href="{{ url('admin/budget-usages/create', $divisi) }}" role="button"><i class="fa fa-plus"></i>&nbsp; {{ trans('admin.budget-usage.actions.create') }}</a>
+                        @can('admin.budget-usage.create')
+                            <a class="btn btn-primary btn-spinner btn-sm pull-right m-b-0" href="{{ url('admin/budget-usages/create', $divisi) }}" role="button"><i class="fa fa-plus"></i>&nbsp; {{ trans('admin.budget-usage.actions.create') }}</a>
+                        @endcan
                     </div>
                     <div class="card-body" v-cloak>
                         <div class="card-block">
@@ -63,12 +65,16 @@
                                         
                                         <td>
                                             <div class="row no-gutters">
-                                                <div class="col-auto">
-                                                    <a class="btn btn-sm btn-spinner btn-info" :href="item.resource_url + '/edit'" title="{{ trans('brackets/admin-ui::admin.btn.edit') }}" role="button"><i class="fa fa-edit"></i></a>
-                                                </div>
-                                                <form class="col" @submit.prevent="deleteItem(item.resource_url)">
-                                                    <button type="submit" class="btn btn-sm btn-danger" title="{{ trans('brackets/admin-ui::admin.btn.delete') }}"><i class="fa fa-trash-o"></i></button>
-                                                </form>
+                                                @can('admin.budget-usage.edit')
+                                                    <div class="col-auto">
+                                                        <a class="btn btn-sm btn-spinner btn-info" :href="item.resource_url + '/edit'" title="{{ trans('brackets/admin-ui::admin.btn.edit') }}" role="button"><i class="fa fa-edit"></i></a>
+                                                    </div>
+                                                @endcan
+                                                @can('admin.budget-usage.delete')
+                                                    <form class="col" @submit.prevent="deleteItem(item.resource_url)">
+                                                        <button type="submit" class="btn btn-sm btn-danger" title="{{ trans('brackets/admin-ui::admin.btn.delete') }}"><i class="fa fa-trash-o"></i></button>
+                                                    </form>
+                                                @endcan
                                             </div>
                                         </td>
                                     </tr>
@@ -88,7 +94,9 @@
                                 <i class="icon-magnifier"></i>
                                 <h3>{{ trans('brackets/admin-ui::admin.index.no_items') }}</h3>
                                 <p>{{ trans('brackets/admin-ui::admin.index.try_changing_items') }}</p>
-                                <a class="btn btn-primary btn-spinner" href="{{ url('admin/budget-usages/create', $divisi) }}" role="button"><i class="fa fa-plus"></i>&nbsp; {{ trans('admin.budget-usage.actions.create') }}</a>
+                                @can('admin.budget-usage.create')
+                                    <a class="btn btn-primary btn-spinner" href="{{ url('admin/budget-usages/create', $divisi) }}" role="button"><i class="fa fa-plus"></i>&nbsp; {{ trans('admin.budget-usage.actions.create') }}</a>
+                                @endcan
                             </div>
                         </div>
                     </div>
